requirements.txt
# NOTE: This requirements.txt is used for local development. All code is now in src/.

html2text
markdown
beautifulsoup4
requests
lxml
tqdm
python-dateutil
feedparser
pytest